refactor test client as reusable between tests
