+++ DevCycle.SDK.Server.Local.MSTests/DevCycleTest.cs
@@ -1,16 +1,9 @@
 ﻿using System;
-using System.Net;
-using System.Threading;
 using System.Threading.Tasks;
 using DevCycle.SDK.Server.Local.Api;
-using DevCycle.SDK.Server.Local.ConfigManager;
-using DevCycle.SDK.Server.Common.API;
 using DevCycle.SDK.Server.Common.Model;
 using DevCycle.SDK.Server.Common.Model.Local;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using RichardSzalay.MockHttp;
 using Environment = System.Environment;
 using System.Collections.Generic;
 using System.Text.Json;
@@ -25,40 +18,6 @@ namespace DevCycle.SDK.Server.Local.MSTests
     [TestClass]
     public class DevCycleTest
     {
-        private DevCycleLocalClient getTestClient(DevCycleLocalOptions options = null, string config = null,
-            bool skipInitialize = false)
-        {
-            config ??= new string(Fixtures.Config());
-
-            var mockHttp = new MockHttpMessageHandler();
-
-            mockHttp.When("https://config-cdn*")
-                .Respond(HttpStatusCode.OK, "application/json",
-                    config);
-            mockHttp.When("https://events*")
-                .Respond(HttpStatusCode.Created, "application/json",
-                    "{}");
-            var localBucketing = new LocalBucketing();
-            var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
-            var sdkKey = $"dvc_server_{Guid.NewGuid().ToString().Replace('-', '_')}_hash";
-            localBucketing.StoreConfig(sdkKey, config);
-            var configManager = new EnvironmentConfigManager(sdkKey, options ?? new DevCycleLocalOptions(),
-                new NullLoggerFactory(),
-                localBucketing,
-                restClientOptions: new DevCycleRestClientOptions() { ConfigureMessageHandler = _ => mockHttp });
-            configManager.Initialized = !skipInitialize;
-
-            DevCycleLocalClient api = new DevCycleLocalClientBuilder()
-                .SetLocalBucketing(localBucketing)
-                .SetConfigManager(configManager)
-                .SetRestClientOptions(new DevCycleRestClientOptions() { ConfigureMessageHandler = _ => mockHttp })
-                .SetOptions(options ?? new DevCycleLocalOptions())
-                .SetSDKKey(sdkKey)
-                .SetLogger(loggerFactory)
-                .Build();
-            return api;
-        }
-
         [TestMethod]
         public async Task CustomCDNURITest()
         {
@@ -108,7 +67,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetFeaturesTest()
         {
-            var api = getTestClient();
+            var api = DevCycleTestClient.getTestClient();
             var user = new DevCycleUser("j_test");
             user.Country = "CA";
             user.Language = "en";
@@ -127,7 +86,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetVariableByKeyTestAsync()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
             var user = new DevCycleUser("j_test");
             string key = "test";
@@ -147,7 +106,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetVariableByKeySpecialCharactersTestAsync()
         {
-            using DevCycleLocalClient api = getTestClient(config: Fixtures.ConfigWithSpecialCharacters());
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient(config: Fixtures.ConfigWithSpecialCharacters());
             var user = new DevCycleUser("j_test");
             string key = Fixtures.VariableKey;
             await Task.Delay(3000);
@@ -168,7 +127,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetVariableByKeyJsonObjTestAsync()
         {
-            using DevCycleLocalClient api = getTestClient(config: Fixtures.ConfigWithJSONValues());
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient(config: Fixtures.ConfigWithJSONValues());
             var user = new DevCycleUser("j_test");
             string key = Fixtures.VariableKey;
             await Task.Delay(3000);
@@ -189,7 +148,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetJsonVariableByKeyReturnsDefaultArrayTest()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
             var user = new DevCycleUser("j_test");
             string key = "json";
@@ -210,7 +169,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetJsonVariableByKeyReturnsDefaultObjectTest()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
             var user = new DevCycleUser("j_test");
             string key = "json";
@@ -231,7 +190,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task GetVariablesTest()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
             DevCycleUser user = new DevCycleUser("j_test");
             // Bucketing needs time to work.
@@ -251,7 +210,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task PostEventsTest()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
 
             DateTimeOffset now = DateTimeOffset.UtcNow;
@@ -267,7 +226,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         {
             Assert.ThrowsException<ArgumentNullException>(() =>
             {
-                using DevCycleLocalClient api = getTestClient();
+                using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
                 var variable = api.Variable(null, "some_key", true).Result;
             });
@@ -293,7 +252,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public void SetClientCustomDataTest()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
 
             Dictionary<string, object> customData = new Dictionary<string, object>();
             customData.Add("strProp", "value");
@@ -351,7 +310,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task TestOpenFeatureInitialization()
         {
-            var dvcClient = getTestClient();
+            var dvcClient = DevCycleTestClient.getTestClient();
             await OpenFeature.Api.Instance.SetProviderAsync(dvcClient.GetOpenFeatureProvider());
             FeatureClient client = OpenFeature.Api.Instance.GetClient();
 
@@ -363,7 +322,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         [TestMethod]
         public async Task TestOpenFeatureJSON()
         {
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             await OpenFeature.Api.Instance.SetProviderAsync(api.GetOpenFeatureProvider());
             FeatureClient client = OpenFeature.Api.Instance.GetClient();
 
@@ -418,7 +377,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         public async Task BeforeHookError_ThrowsException()
         {
             const string key = "test";
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             TestEvalHook hook = new TestEvalHook() { ThrowBefore = true };
             api.AddEvalHook(hook);
 
@@ -439,7 +398,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         public async Task AfterHookError_ThrowsException()
         {
             const string key = "test";
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             TestEvalHook hook = new TestEvalHook() { ThrowAfter = true };
             api.AddEvalHook(hook);
 
@@ -461,7 +420,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         public async Task ErrorHookError_ThrowsException()
         {
             const string key = "test";
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             TestEvalHook hook = new TestEvalHook() { ThrowError = true, ThrowAfter = true };
             api.AddEvalHook(hook);
 
@@ -483,7 +442,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         public async Task FinallyHookError_ThrowsException()
         {
             const string key = "test";
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             TestEvalHook hook = new TestEvalHook() { ThrowFinally = true };
             api.AddEvalHook(hook);
 
@@ -505,7 +464,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         public async Task EvalHooks_NormalExecution()
         {
             const string key = "test";
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             TestEvalHook hook = new TestEvalHook();
             api.AddEvalHook(hook);
 
@@ -527,7 +486,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
         public async Task ClearEvalHooks_RemovesAllHooks()
         {
             const string key = "test";
-            using DevCycleLocalClient api = getTestClient();
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
             TestEvalHook hook1 = new TestEvalHook();
             TestEvalHook hook2 = new TestEvalHook();
             api.AddEvalHook(hook1);
@@ -557,7 +516,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
             {
                 EvalHooks = [hook]
             };
-            using DevCycleLocalClient api = getTestClient(options);
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient(options);
 
             await Task.Delay(3000);
             var result = await api.VariableAsync(new DevCycleUser("test"), key, true);
@@ -583,7 +542,7 @@ namespace DevCycle.SDK.Server.Local.MSTests
             {
                 EvalHooks = [hook1, hook2]
             };
-            using DevCycleLocalClient api = getTestClient(options);
+            using DevCycleLocalClient api = DevCycleTestClient.getTestClient(options);
 
             await Task.Delay(3000);
             var result = await api.VariableAsync(new DevCycleUser("test"), key, true);
