+++ DevCycle.SDK.Server.Local.MSTests/EvalReason.cs
@@ -1,62 +1,19 @@
-using System;
-using System.Net;
 using System.Threading.Tasks;
 using DevCycle.SDK.Server.Local.Api;
-using DevCycle.SDK.Server.Local.ConfigManager;
-using DevCycle.SDK.Server.Common.API;
 using DevCycle.SDK.Server.Common.Model;
-using DevCycle.SDK.Server.Common.Model.Local;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
-using RichardSzalay.MockHttp;
 
 namespace DevCycle.SDK.Server.Local.MSTests;
 
 [TestClass]
 public class EvalReasonTests
 {
-    private DevCycleLocalClient getTestClient(DevCycleLocalOptions options = null, string config = null,
-        bool skipInitialize = false)
-    {
-        config ??= new string(Fixtures.Config());
-
-        var mockHttp = new MockHttpMessageHandler();
-
-        mockHttp.When("https://config-cdn*")
-            .Respond(skipInitialize ? HttpStatusCode.BadRequest : HttpStatusCode.OK, "application/json",
-                config);
-        mockHttp.When("https://events*")
-            .Respond(HttpStatusCode.Created, "application/json",
-                "{}");
-        var localBucketing = new LocalBucketing();
-        var loggerFactory = LoggerFactory.Create(builder => builder.AddConsole());
-        var sdkKey = $"dvc_server_{Guid.NewGuid().ToString().Replace('-', '_')}_hash";
-        localBucketing.StoreConfig(sdkKey, config);
-        var configManager = new EnvironmentConfigManager(sdkKey, options ?? new DevCycleLocalOptions(),
-            new NullLoggerFactory(),
-            localBucketing,
-            restClientOptions: new DevCycleRestClientOptions() { ConfigureMessageHandler = _ => mockHttp });
-        configManager.Initialized = !skipInitialize;
-
-        DevCycleLocalClient api = new DevCycleLocalClientBuilder()
-            .SetLocalBucketing(localBucketing)
-            .SetConfigManager(configManager)
-            .SetRestClientOptions(new DevCycleRestClientOptions() { ConfigureMessageHandler = _ => mockHttp })
-            .SetOptions(options ?? new DevCycleLocalOptions())
-            .SetSDKKey(sdkKey)
-            .SetLogger(loggerFactory)
-            .Build();
-        return api;
-    }
-
     // ===== Tests for Default Reasons - Variable Method =====
-
     [TestMethod]
     public async Task Variable_MissingConfig_ReturnsDefaultWithMissingConfigReason()
     {
-        using DevCycleLocalClient api = getTestClient(skipInitialize: true);
+        using DevCycleLocalClient api = DevCycleTestClient.getTestClient(skipInitialize: true);
         var user = new DevCycleUser("test_user");
         const string key = "test_variable";
         const bool defaultValue = false;
@@ -72,7 +29,7 @@ public class EvalReasonTests
     [TestMethod]
     public async Task Variable_UserNotTargeted_ReturnsDefaultWithUserNotTargetedReason()
     {
-        using DevCycleLocalClient api = getTestClient();
+        using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
         var user = new DevCycleUser("test_user");
         const string key = "non_existent_variable";
         const string defaultValue = "default_string";
@@ -90,7 +47,7 @@ public class EvalReasonTests
     [TestMethod]
     public async Task VariableAsync_MissingConfig_ReturnsDefaultWithMissingConfigReason()
     {
-        using DevCycleLocalClient api = getTestClient(skipInitialize: true);
+        using DevCycleLocalClient api = DevCycleTestClient.getTestClient(skipInitialize: true);
         var user = new DevCycleUser("test_user");
         const string key = "test_variable";
         const int defaultValue = 42;
@@ -106,7 +63,7 @@ public class EvalReasonTests
     [TestMethod]
     public async Task VariableAsync_UserNotTargeted_ReturnsDefaultWithUserNotTargetedReason()
     {
-        using DevCycleLocalClient api = getTestClient();
+        using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
         var user = new DevCycleUser("test_user");
         const string key = "non_existent_variable";
         const double defaultValue = 3.14;
@@ -123,7 +80,7 @@ public class EvalReasonTests
     [TestMethod]
     public async Task Variable_Bool_MismatchType_ReturnsDefault()
     {
-        using DevCycleLocalClient api = getTestClient();
+        using DevCycleLocalClient api = DevCycleTestClient.getTestClient();
         var user = new DevCycleUser("test_user");
         const string key = "test";
         const bool validDefaultValue = false;
@@ -141,7 +98,7 @@ public class EvalReasonTests
     [TestMethod]
     public async Task Variable_JSON_MismatchType_ReturnsDefault()
     {
-        using DevCycleLocalClient api = getTestClient(config: Fixtures.ConfigWithJSONValues());
+        using DevCycleLocalClient api = DevCycleTestClient.getTestClient(config: Fixtures.ConfigWithJSONValues());
         var user = new DevCycleUser("test_user");
         const string key = "test";
         var validDefaultValue = Newtonsoft.Json.Linq.JObject.Parse("{\"key\": \"default\"}");
